charity/app/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
django-menu-generator==1.1.0
gunicorn==20.0.4
idna==2.10
packaging==20.9
psycopg2==2.8.6
PyMySQL==1.0.2
pyparsing==2.4.7
pypi-simple==0.8.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
vulners==1.5.11
whitenoise==5.2.0
django-paypal
paynow